move default service types list to types
