+++ config/types.js
@@ -133,3 +133,26 @@ global:
 };
 
 export const SYSTEM_PROJECT_LABEL = 'authz.management.cattle.io/system-project';
+
+export const DEFAULT_SERVICE_TYPES = [
+  {
+    id:               'ClusterIP',
+    translationLabel: 'serviceTypes.clusterip'
+  },
+  {
+    id:               'ExternalName',
+    translationLabel: 'serviceTypes.externalname'
+  },
+  {
+    id:               'Headless',
+    translationLabel: 'serviceTypes.headless'
+  },
+  {
+    id:               'LoadBalancer',
+    translationLabel: 'serviceTypes.loadbalancer'
+  },
+  {
+    id:               'NodePort',
+    translationLabel: 'serviceTypes.nodeport'
+  },
+];
